Move ContentArea into base
