+++ lib/action_view/component.rb
@@ -13,7 +13,6 @@ module ActionView
     autoload :Conversion
     autoload :Preview
     autoload :Previewable
-    autoload :ContentAreas
     autoload :TestHelpers
     autoload :TestCase
     autoload :RenderMonkeyPatch
